+++ Sources/MapboxCoreNavigation/Feedback/EndOfRouteFeedback.swift
@@ -7,12 +7,12 @@ open class EndOfRouteFeedback {
     /**
      Rating: The user's rating for the route. Normalized between 0 and 100.
      */
-    let rating: Int?
+    public let rating: Int?
     
     /**
      Comment: Any comments that the user had about the route.
      */
-    let comment: String?
+    public let comment: String?
     
     @nonobjc public init(rating: Int? = nil, comment: String? = nil) {
         self.rating = rating